requirements.txt
altair==5.5.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
logzero==1.7.0
MarkupSafe==3.0.2
narwhals==1.44.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyotp==2.9.0
pyspark==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smartapi-python==1.5.5
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.0
ta==0.11.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0 